startup.cs update

+++ RVASIspit/Controllers/ZaposleniController.cs
@@ -18,11 +18,21 @@ namespace RVASIspit.Controllers
             db.Dispose();
         }
 
+        // GET: Zaposleni
+        [Authorize(Roles = "Admin, Korisnik")]
         public ActionResult Index()
         {
-            return View(db.Zaposleni.ToList());
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
+            {
+                return View(db.Zaposleni.ToList());
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
         }
 
+        // GET: Zaposleni/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -37,25 +47,41 @@ namespace RVASIspit.Controllers
             return View(zaposleni);
         }
 
+        // GET: Zaposleni/Create
         public ActionResult Create()
         {
-            return View();
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
+            {
+                return View();
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
 
+        // POST: Zaposleni/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ZaposleniID,Ime,Prezime")] Zaposleni zaposleni)
         {
-            if (ModelState.IsValid)
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
             {
-                db.Zaposleni.Add(zaposleni);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Zaposleni.Add(zaposleni);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(zaposleni);
             }
-
-            return View(zaposleni);
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
 
+        // GET: Zaposleni/Edit/5
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -67,22 +93,38 @@ namespace RVASIspit.Controllers
             {
                 return HttpNotFound();
             }
-            return View(zaposleni);
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
+            {
+                return View(zaposleni);
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
 
+        // POST: Zaposleni/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ZaposleniID,Ime,Prezime")] Zaposleni zaposleni)
         {
-            if (ModelState.IsValid)
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
             {
-                db.Entry(zaposleni).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(zaposleni).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(zaposleni);
             }
-            return View(zaposleni);
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
 
+        // GET: Zaposleni/Delete/5
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -94,17 +136,32 @@ namespace RVASIspit.Controllers
             {
                 return HttpNotFound();
             }
-            return View(zaposleni);
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
+            {
+                return View(zaposleni);
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
 
+        // POST: Zaposleni/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Zaposleni zaposleni = db.Zaposleni.Find(id);
-            db.Zaposleni.Remove(zaposleni);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (User.IsInRole("Admin") || User.IsInRole("Korisnik"))
+            {
+                Zaposleni zaposleni = db.Zaposleni.Find(id);
+                db.Zaposleni.Remove(zaposleni);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden); // Return 403 Forbidden for unauthorized users
+            }
         }
     }
 }
